requirements.txt
cadquery>=2.7.0
pygcode>=0.2.1
numpy>=2.0
scipy>=1.14
matplotlib>=3.10
pyvista>=0.47
pyvistaqt>=0.11
PySide6>=6.7
pyyaml>=6.0
# Optional CNC connections (install as needed):
# pyserial>=3.5        # For GRBL serial connection
# opcua>=0.98          # For OPC UA protocol
# requests>=2.31       # For MTConnect HTTP (already in most envs)
